+++ Tools/c-analyzer/cpython/globals-to-fix.tsv
@@ -303,20 +303,9 @@ Objects/sliceobject.c	-	_Py_EllipsisObject	-
 ## global non-objects to fix in core code
 
 ##-----------------------
-## effectively-const but initialized lazily
-
-## others
-Python/perf_trampoline.c	-	perf_map_file	-
-
-##-----------------------
 ## state
 
-## other
 Objects/object.c	-	_Py_RefTotal	-
-Python/perf_trampoline.c	-	perf_status	-
-Python/perf_trampoline.c	-	extra_code_index	-
-Python/perf_trampoline.c	-	code_arena	-
-Python/perf_trampoline.c	-	trampoline_api	-
 Python/thread_pthread_stubs.h	-	py_tls_entries	-
 
 
@@ -374,9 +363,8 @@ Modules/itertoolsmodule.c	-	teedataobject_type	-
 Modules/itertoolsmodule.c	-	ziplongest_type	-
 
 ##-----------------------
-## other
-
 ## state
+
 Modules/faulthandler.c	-	fatal_error	-
 Modules/faulthandler.c	-	thread	-
 Modules/faulthandler.c	-	user_signals	-
